wolf3D

+++ wolf3D/mlx_linux/include/colors.h
@@ -6,7 +6,7 @@
 /*   By: wealdboar <marvin@42.fr>                   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/08/14 03:30:25 by wealdboar         #+#    #+#             */
-/*   Updated: 2020/08/22 11:42:02 by lchantel         ###   ########.fr       */
+/*   Updated: 2020/09/18 03:44:22 by lchantel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 
 typedef struct			colors
 {
-	int					trnsp;
-	int					red;
-	int					green;
-	int					blue;
-	unsigned int		_clrfull;
+	unsigned long		trnsp;
+	unsigned long		red;
+	unsigned long		green;
+	unsigned long		blue;
+	unsigned long		_clrfull;
 }						bitmap;
 
 int		get_bitmaps(bitmap vals);
@@ -28,7 +28,7 @@ int		get_rcode(bitmap vals);
 int		get_gcode(bitmap vals);
 int		get_bcode(bitmap vals);
 bitmap	add_shade(double dist, bitmap vals);
-void	set_color(bitmap *clr, int trnsp, int red,
-		int green, int blue);
+void	set_color(bitmap *clr, unsigned long trnsp, unsigned long red,
+		unsigned long green, unsigned long blue);
 
 #endif
